move edm4hep tests into one file
